Issue #ED-3836: Angular-11 upgrade
