+++ demo/src/app/app.routing.ts
@@ -128,7 +128,7 @@ const appRoutes:Routes = [
 
 @NgModule({
     imports: [
-        RouterModule.forRoot(appRoutes, { useHash: true })
+        RouterModule.forRoot(appRoutes, { useHash: true, relativeLinkResolution: 'legacy' })
     ],
     exports: [
         RouterModule
