+++ package.json
@@ -43,31 +43,30 @@
     "build:prod": "npm run lib:build:prod && npm run demo:build:prod"
   },
   "dependencies": {
-    "bowser": "^1.7.2",
-    "date-fns": "2.0.0-alpha.1",
-    "element-closest": "^2.0.2",
-    "is-ua-webview": "^1.0.0",
-    "popper.js": "^1.14.0",
-    "tslib": "^2.0.0"
+    "bowser": "2.11.0",
+    "date-fns": "3.6.0",
+    "element-closest": "3.0.2",
+    "is-ua-webview": "1.1.2",
+    "popper.js": "^1.16.1",
+    "tslib": "^2.6.2"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "0.1102.19",
-    "@angular-devkit/build-ng-packagr": "~0.1002.4",
-    "@angular-devkit/core":"10.2.1",
-    "@angular/animations": "^10.2.5",
-    "@angular/cli": "~10.2.4",
-    "@angular/common": "^10.2.5",
-    "@angular/compiler": "^10.2.5",
-    "@angular/compiler-cli": "^10.2.5",
-    "@angular/core": "^10.2.5",
-    "@angular/forms": "^10.2.5",
-    "@angular/language-service": "^10.2.5",
-    "@angular/platform-browser": "^10.2.5",
-    "@angular/platform-browser-dynamic": "^10.2.5",
-    "@angular/router": "^10.2.5",
+    "@angular-devkit/core": "11.2.19",
+    "@angular/animations": "^11.2.14",
+    "@angular/cli": "~11.2.19",
+    "@angular/common": "^11.2.14",
+    "@angular/compiler": "^11.2.14",
+    "@angular/compiler-cli": "^11.2.14",
+    "@angular/core": "^11.2.14",
+    "@angular/forms": "^11.2.14",
+    "@angular/language-service": "^11.2.14",
+    "@angular/platform-browser": "^11.2.14",
+    "@angular/platform-browser-dynamic": "^11.2.14",
+    "@angular/router": "^11.2.14",
     "@types/extend": "3.0.0",
-    "@types/jasmine": "2.8.6",
-    "@types/jasminewd2": "2.0.3",
+    "@types/jasmine": "~3.6.0",
+    "@types/jasminewd2": "2.0.13",
     "@types/node": "^12.11.1",
     "@types/prismjs": "1.4.18",
     "bowser": "^1.7.2",
@@ -79,7 +78,7 @@
     "is-ua-webview": "^1.0.0",
     "jasmine-core": "3.8.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.4.3",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
@@ -87,7 +86,7 @@
     "karma-min-reporter": "^0.1.0",
     "karma-phantomjs-launcher": "^1.0.4",
     "karma-remap-istanbul": "~0.4.0",
-    "ng-packagr": "^10.1.0",
+    "ng-packagr": "^11.2.4",
     "popper.js": "^1.14.0",
     "prismjs": "~1.6.0",
     "rxjs": "6.6.7",
