+++ src/misc/util/helpers/util.ts
@@ -16,6 +16,10 @@ interface IRecursiveObject { [name:string]:IRecursiveObject; }
 
 export interface ITemplateRefContext<T> { $implicit:T; }
 
+export interface IResultContext<T> extends ITemplateRefContext<T> {
+    query:string;
+}
+
 export interface IAugmentedElement extends Element {
     closest(selector:string):IAugmentedElement;
 }
