+++ src/modules/datepicker/classes/datepicker-mode.ts
@@ -1,0 +1,9 @@
+export type DatepickerMode = "year" | "month" | "date" | "datetime" | "time";
+
+export const DatepickerMode = {
+    Year: "year" as DatepickerMode,
+    Month: "month" as DatepickerMode,
+    Date: "date" as DatepickerMode,
+    Datetime: "datetime" as DatepickerMode,
+    Time: "time" as DatepickerMode
+};
