+++ src/modules/datepicker/components/datepicker.ts
@@ -3,16 +3,6 @@ import { CalendarService } from "./../services/calendar.service";
 import { DatetimeConfig } from "../classes/calendar-config";
 import { SuiLocalizationService } from "../../../behaviors/localization/internal";
 
-export type DatepickerMode = "year" | "month" | "date" | "datetime" | "time";
-
-export const DatepickerMode = {
-    Year: "year" as DatepickerMode,
-    Month: "month" as DatepickerMode,
-    Date: "date" as DatepickerMode,
-    Datetime: "datetime" as DatepickerMode,
-    Time: "time" as DatepickerMode
-};
-
 @Component({
     selector: "sui-datepicker",
     template: `
