+++ src/tsconfig.json
@@ -2,6 +2,7 @@
   "extends": "../tsconfig.json",
   "compilerOptions": {
     "outDir": "../temp",
+    "declarationMap": true,
     "target": "es2015",
     "module": "es2015",
     "moduleResolution": "node",
